refactor(routes): extract jellyfin catalog routes
